[actor] Migrate PartialWitnessActor to use ActixWrapper
