+++ chain/client/src/stateless_validation/partial_witness/mod.rs
@@ -1,3 +1,2 @@
-pub mod partial_witness_actions;
 pub mod partial_witness_actor;
 mod partial_witness_tracker;
