+++ chain/client/src/test_utils/test_loop.rs
@@ -1,5 +1,5 @@
 pub mod client_actions;
-pub mod partial_witness_actions;
+pub mod partial_witness_actor;
 pub mod sync_actor;
 pub mod sync_jobs_actions;
